refactor: split out native and web prop types for ad height
