+++ packages/ad/src/dom-context-prop-types.js
@@ -1,17 +1,3 @@
-import PropTypes from "prop-types";
+import { propTypes, defaultProps } from "./dom-context-prop-types-base";
 
-export const propTypes = {
-  data: PropTypes.shape({}),
-  height: PropTypes.number,
-  init: PropTypes.func.isRequired,
-  onRenderComplete: PropTypes.func,
-  onRenderError: PropTypes.func,
-  width: PropTypes.number.isRequired
-};
-
-export const defaultProps = {
-  data: {},
-  height: 0,
-  onRenderComplete: () => {},
-  onRenderError: () => {}
-};
+export { propTypes, defaultProps };
